feat!: rewrite lodkit.ttl constructor
The change introduces an overhaul and partial rewrite of the
lodkit.ttl constructor.

1. Discontinuation of skip functionality and graph parameter in the initializer

The skip functionality is an awkward and ineffecient feature that
isn't really needed, so it might as well be cut from the implementation.
Also, passing a graph instance to the initializer introduces state and
is not in line with the idea of the to_graph method as a mere
convenience method that accepts a graph instance anyways.

2. Making lodkit.ttl an Iterable

lodkit.ttl should be an Iterable, not an Iterator. Functionally, the
constructor could also be expressed as a generator function (see the
internal lodkit.ttl._generate_triples); using a class however allows
to match/case against the lodkit.ttl type and express a union type of
_TPredicateObjectPair *recursively*.

Closes #26.
Closes #27.
Closes #28.