feat!: implement simplified URIConstructor
The change discontinues the old URIConstructorFactory in favor of a
simplified URIConstructor class.

URIConstructor takes a namespaces and constructs URIRefs on __call__.

If the hash_value argument is of type str | bytes, a URI with a
hashed component is produced for the given namespace. Note that the
old URIConstructorFactory was not able to generate hashes based on bytes.

If hash_value is None (the default), a URI with a unique component for the
given namespace is generated using UUID4.